feat(service): add business rules validation

+++ internal/product/service.go
@@ -1,5 +1,10 @@
 package product
 
+import (
+	"fmt"
+	"strings"
+)
+
 type productService struct {
 	repo *productRepository
 }
@@ -15,13 +20,25 @@ func (s *productService) GetAllProducts() ([]Product, error) {
 }
 
 func (s *productService) GetProductByID(id int) (Product, error) {
+	if id <= 0 {
+		return Product{}, fmt.Errorf("invalid id")
+	}
 	return s.repo.GetByID(id)
 }
 
 func (s *productService) CreateProduct(product Product) (Product, error) {
+	if strings.TrimSpace(product.Name) == "" {
+		return Product{}, fmt.Errorf("invalid name")
+	}
+	if product.Price < 0 {
+		return Product{}, fmt.Errorf("invalid price")
+	}
 	return s.repo.Create(product)
 }
 
 func (s *productService) DeleteProduct(id int) error {
+	if id <= 0 {
+		return fmt.Errorf("invalid id")
+	}
 	return s.repo.Delete(id)
 }
